Simplify HUPPOrderIssueProducer by extracting create-draft-issues code
Issues HUs are still selectable in HU editor while production is not
completed
metasfresh/metasfresh-webui-api#590
