+++ de.metas.handlingunits.base/src/main/java/de/metas/handlingunits/pporder/api/impl/HUPPOrderIssueProducer.java
@@ -1,7 +1,5 @@
 package de.metas.handlingunits.pporder.api.impl;
 
-import static org.adempiere.model.InterfaceWrapperHelper.newInstance;
-
 /*
  * #%L
  * de.metas.handlingunits.base
@@ -28,41 +26,31 @@ import java.util.Collection;
 import java.util.Date;
 import java.util.List;
 
-import javax.annotation.Nonnull;
-
 import org.adempiere.model.IContextAware;
 import org.adempiere.model.InterfaceWrapperHelper;
 import org.adempiere.util.Check;
 import org.adempiere.util.GuavaCollectors;
 import org.adempiere.util.Services;
-import org.adempiere.util.time.SystemTime;
-import org.compiere.util.TimeUtil;
 import org.eevolution.api.IPPOrderBOMDAO;
 import org.eevolution.model.I_PP_Order;
 import org.eevolution.model.I_PP_Order_BOMLine;
-import org.eevolution.model.X_PP_Order_BOMLine;
-import org.slf4j.Logger;
 
+import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.MoreObjects;
 import com.google.common.collect.ImmutableList;
 
 import de.metas.handlingunits.IHUContext;
 import de.metas.handlingunits.IHandlingUnitsBL;
 import de.metas.handlingunits.IHandlingUnitsDAO;
-import de.metas.handlingunits.attribute.IPPOrderProductAttributeBL;
 import de.metas.handlingunits.attribute.IPPOrderProductAttributeDAO;
 import de.metas.handlingunits.exceptions.HUException;
-import de.metas.handlingunits.hutransaction.IHUTrxBL;
 import de.metas.handlingunits.model.I_M_HU;
 import de.metas.handlingunits.model.I_PP_Order_Qty;
 import de.metas.handlingunits.model.X_M_HU;
 import de.metas.handlingunits.pporder.api.IHUPPOrderIssueProducer;
 import de.metas.handlingunits.pporder.api.IHUPPOrderQtyDAO;
-import de.metas.handlingunits.storage.IHUProductStorage;
-import de.metas.logging.LogManager;
-import de.metas.material.planning.pporder.IPPOrderBOMBL;
+import de.metas.handlingunits.pporder.api.impl.hu_pporder_issue_producer.CreateDraftIssues;
 import de.metas.material.planning.pporder.PPOrderUtil;
-import de.metas.quantity.Quantity;
 import lombok.NonNull;
 
 /**
@@ -73,27 +61,16 @@ import lombok.NonNull;
  */
 public class HUPPOrderIssueProducer implements IHUPPOrderIssueProducer
 {
-	// Services
-	private static final transient Logger logger = LogManager.getLogger(HUPPOrderIssueProducer.class);
-	//
-	private final transient IPPOrderBOMBL ppOrderBOMBL = Services.get(IPPOrderBOMBL.class);
-	//
-	private final transient IHUTrxBL huTrxBL = Services.get(IHUTrxBL.class);
 	private final transient IHandlingUnitsBL handlingUnitsBL = Services.get(IHandlingUnitsBL.class);
 	private final transient IHUPPOrderQtyDAO huPPOrderQtyDAO = Services.get(IHUPPOrderQtyDAO.class);
-	private final transient IPPOrderProductAttributeBL ppOrderProductAttributeBL = Services.get(IPPOrderProductAttributeBL.class);
 	private final transient IPPOrderProductAttributeDAO ppOrderProductAttributeDAO = Services.get(IPPOrderProductAttributeDAO.class);
-
-	private static final String MSG_IssuingAggregatedTUsNotAllowed = "de.metas.handlingunits.pporder.api.impl.HUPPOrderIssueProducer.IssuingAggregatedTUsNotAllowed";
-	private static final String MSG_IssuingVHUsNotAllowed = "de.metas.handlingunits.pporder.api.impl.HUPPOrderIssueProducer.IssuingVHUsNotAllowed";
-	private static final String MSG_IssuingHUWithMultipleProductsNotAllowed = "de.metas.handlingunits.pporder.api.impl.HUPPOrderIssueProducer.IssuingHUsWithMultipleProductsNotAllowed";
 
 	private Date movementDate;
 	private List<I_PP_Order_BOMLine> targetOrderBOMLines;
 
+	@VisibleForTesting
 	HUPPOrderIssueProducer()
 	{
-		super();
 	}
 
 	@Override
@@ -107,176 +84,14 @@ public class HUPPOrderIssueProducer implements IHUPPOrderIssueProducer
 	}
 
 	@Override
-	public List<I_PP_Order_Qty> createDraftIssues(@Nonnull final Collection<I_M_HU> hus)
-	{
-		if (hus.isEmpty())
-		{
-			return ImmutableList.of();
-		}
-
-		// NOTE: we would prefer to always run this out of transactions,
-		// but in some cases like issuing from Swing POS we cannot enforce it because in that case
-		// the candidates are created and processed in one uber-transaction
-		// trxManager.assertThreadInheritedTrxNotExists();
-
-		final List<I_PP_Order_Qty> candidates = huTrxBL.process(huContext -> {
-			return hus.stream()
-					.map(hu -> createIssue_InTrx(huContext, hu))
-					.filter(issueCandidate -> issueCandidate != null)
-					.collect(ImmutableList.toImmutableList());
-		});
-		return candidates;
-	}
-
-	private I_PP_Order_Qty createIssue_InTrx(
-			@NonNull final IHUContext huContext,
-			@NonNull final I_M_HU hu)
-	{
-		if (!X_M_HU.HUSTATUS_Active.equals(hu.getHUStatus()))
-		{
-			throw new HUException("Only active HUs can be issued but " + hu + " is " + hu.getHUStatus());
-		}
-
-		removeHuFromParent(huContext, hu);
-
-		final IHUProductStorage productStorage = retrieveProductStorage(huContext, hu);
-		if (productStorage == null)
-		{
-			return null;
-		}
-
-		// Actually create and save the candidate
-		final I_PP_Order_Qty candidate = createIssueCandidate(hu, productStorage);
-
-		// update the HU's status so that it's not moved somewhere else etc
-		handlingUnitsBL.setHUStatus(huContext, hu, X_M_HU.HUSTATUS_Issued);
-		Services.get(IHandlingUnitsDAO.class).saveHU(hu);
-
-		return candidate;
-	}
-
-	/**
-	 * If not a top level HU, take it out first
-	 * 
-	 * @param huContext
-	 * @param hu
-	 */
-	private void removeHuFromParent(
-			@NonNull final IHUContext huContext,
-			@NonNull final I_M_HU hu)
-	{
-		if (!handlingUnitsBL.isTopLevel(hu))
-		{
-			if (handlingUnitsBL.isAggregateHU(hu))
-			{
-				throw HUException.ofAD_Message(MSG_IssuingAggregatedTUsNotAllowed);
-			}
-			if (handlingUnitsBL.isVirtual(hu))
-			{
-				throw HUException.ofAD_Message(MSG_IssuingVHUsNotAllowed);
-			}
-			else
-			{
-				huTrxBL.setParentHU(huContext //
-						, null // parentHUItem
-						, hu //
-						, true // destroyOldParentIfEmptyStorage
-				);
-			}
-		}
-	}
-
-	private IHUProductStorage retrieveProductStorage(
-			@NonNull final IHUContext huContext,
-			@NonNull final I_M_HU hu)
-	{
-		final List<IHUProductStorage> productStorages = huContext.getHUStorageFactory()
-				.getStorage(hu)
-				.getProductStorages();
-
-		// Empty HU
-		if (productStorages.isEmpty())
-		{
-			logger.warn("{}: Skip {} from issuing because its storage is empty", this, hu);
-			return null; // no candidate
-		}
-
-		if (productStorages.size() != 1)
-		{
-			throw HUException.ofAD_Message(MSG_IssuingHUWithMultipleProductsNotAllowed)
-					.setParameter("HU", hu)
-					.setParameter("ProductStorages", productStorages);
-		}
-		final IHUProductStorage productStorage = productStorages.get(0);
-		return productStorage;
-	}
-
-	private I_PP_Order_Qty createIssueCandidate(
-			@NonNull final I_M_HU hu,
-			@NonNull final IHUProductStorage productStorage)
-	{
-		final int productId = productStorage.getM_Product_ID();
-		final I_PP_Order_BOMLine targetBOMLine = getTargetOrderBOMLine(productId);
-
-		final I_PP_Order_Qty candidate = newInstance(I_PP_Order_Qty.class);
-		
-		candidate.setPP_Order_ID(targetBOMLine.getPP_Order_ID());
-		candidate.setPP_Order_BOMLine(targetBOMLine);
-
-		candidate.setM_Locator_ID(hu.getM_Locator_ID());
-		candidate.setM_HU_ID(hu.getM_HU_ID());
-		candidate.setM_Product_ID(productId);
-
-		final Quantity qtyToIssue = calculateQtyToIssue(targetBOMLine, productStorage);
-		candidate.setQty(qtyToIssue.getQty());
-		candidate.setC_UOM(qtyToIssue.getUOM());
-
-		candidate.setMovementDate(TimeUtil.asTimestamp(getMovementDate()));
-		candidate.setProcessed(false);
-		huPPOrderQtyDAO.save(candidate);
-
-		ppOrderProductAttributeBL.addPPOrderProductAttributesFromIssueCandidate(candidate);
-		return candidate;
-	}
-
-	/** @return how much quantity to take "from" and issue it to given BOM line */
-	private Quantity calculateQtyToIssue(final I_PP_Order_BOMLine targetBOMLine, final IHUProductStorage from)
-	{
-		//
-		// Case: if this is an Issue BOM Line, IssueMethod is Backflush and we did not over-issue on it yet
-		// => enforce the capacity to Projected Qty Required (i.e. standard Qty that needs to be issued on this line).
-		// initial concept: http://dewiki908/mediawiki/index.php/07433_Folie_Zuteilung_Produktion_Fertigstellung_POS_%28102170996938%29
-		// additional (use of projected qty required): http://dewiki908/mediawiki/index.php/07601_Calculation_of_Folie_in_Action_Receipt_%28102017845369%29
-		final String issueMethod = targetBOMLine.getIssueMethod();
-		if (X_PP_Order_BOMLine.ISSUEMETHOD_IssueOnlyForReceived.equals(issueMethod))
-		{
-			return ppOrderBOMBL.calculateQtyToIssueBasedOnFinishedGoodReceipt(targetBOMLine, from.getC_UOM());
-		}
-		else
-		{
-			return Quantity.of(from.getQty(), from.getC_UOM());
-		}
-
-	}
-
-	@Override
 	public IHUPPOrderIssueProducer setMovementDate(final Date movementDate)
 	{
 		this.movementDate = movementDate;
 		return this;
 	}
 
-	private Date getMovementDate()
-	{
-		if (movementDate == null)
-		{
-			movementDate = SystemTime.asDayTimestamp();
-		}
-		return movementDate;
-	}
-
 	@Override
-	public IHUPPOrderIssueProducer setTargetOrderBOMLines(final List<I_PP_Order_BOMLine> targetOrderBOMLines)
+	public IHUPPOrderIssueProducer setTargetOrderBOMLines(@NonNull final List<I_PP_Order_BOMLine> targetOrderBOMLines)
 	{
 		Check.assumeNotEmpty(targetOrderBOMLines, "Parameter targetOrderBOMLines is not empty");
 		targetOrderBOMLines.forEach(bomLine -> {
@@ -290,20 +105,9 @@ public class HUPPOrderIssueProducer implements IHUPPOrderIssueProducer
 		return this;
 	}
 
-	private List<I_PP_Order_BOMLine> getTargetOrderBOMLines()
-	{
-		if (targetOrderBOMLines == null || targetOrderBOMLines.isEmpty())
-		{
-			throw new HUException("No BOM lines were configured");
-		}
-
-		return targetOrderBOMLines;
-	}
-
 	@Override
-	public IHUPPOrderIssueProducer setTargetOrderBOMLine(final I_PP_Order_BOMLine targetOrderBOMLine)
+	public IHUPPOrderIssueProducer setTargetOrderBOMLine(@NonNull final I_PP_Order_BOMLine targetOrderBOMLine)
 	{
-		Check.assumeNotNull(targetOrderBOMLine, "targetOrderBOMLine not null");
 		return setTargetOrderBOMLines(ImmutableList.of(targetOrderBOMLine));
 	}
 
@@ -323,16 +127,13 @@ public class HUPPOrderIssueProducer implements IHUPPOrderIssueProducer
 		return setTargetOrderBOMLines(ppOrderBOMLines);
 	}
 
-	private I_PP_Order_BOMLine getTargetOrderBOMLine(final int productId)
-	{
-		final List<I_PP_Order_BOMLine> targetBOMLines = getTargetOrderBOMLines();
-		return targetBOMLines
-				.stream()
-				.filter(bomLine -> bomLine.getM_Product_ID() == productId)
-				.findFirst()
-				.orElseThrow(() -> new HUException("No BOM line found for productId=" + productId + " in " + targetBOMLines));
-	}
 
+	@Override
+	public List<I_PP_Order_Qty> createDraftIssues(@NonNull final Collection<I_M_HU> hus)
+	{
+		return new CreateDraftIssues(targetOrderBOMLines, movementDate).createDraftIssues(hus);
+	}
+	
 	@Override
 	public void reverseDraftIssue(@NonNull final I_PP_Order_Qty candidate)
 	{
